Created routes, moved camera to component, updated JSX in login,deleted unused html files
