--- client/public/camera.html
@@ -1,36 +1,0 @@
-<!DOCTYPE <!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Card Reader</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.3/css/all.css">
-    <link rel="stylesheet" type="text/css" media="screen" href="../src/index.css">
-</head>
-
-<body>
-    <div class="camera-page">
-        <h1>Camera</h1>
-        <div class="form">
-            <form class="login-form">
-                <br />
-                <h4>Capture Your Business Card</h4>
-                <video id="player" width=300 height=300 autoplay class="display"></video>
-                <canvas id="canvas" width=300 height=300 class="hidden"></canvas><br>
-                <button class="mt-3" id="capture">Take Picture</button>
-            </form>
-        </div>
-    </div>
-
-    <script src="https://code.jquery.com/jquery-3.3.1.min.js"
-    integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>
-   
-    <script type="text/javascript" src="./main.js"></script>
-
-</body>
-
-</html>
