--- client/public/login.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <link href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet"
-        integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN" crossorigin="anonymous">
-    <link rel="stylesheet" href="./css/style.css">
-
-    <title>CaptureMe Login Page</title>
-</head>
-
-<body>
-    <div class="login-page" style="height: 100vh;">
-        <div role="heading">
-            <h1>Log In</h1>
-        </div>
-        <div class="form" role="form">
-            <form class="login-form">
-                <input type="text" placeholder="username" />
-                <input type="text" placeholder="password" />
-                <button>login</button>
-                <p class="message">Not registered? <a href="create-login.html"><strong>Sign Up</strong></a></p>
-            </form>
-        </div>
-    </div>
-</body>
-
-</html>
