+++ client/src/components/login/login.js
@@ -1,25 +1,43 @@
-import React from "react";
+import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
 import "./login.css";
 
-var login = React.createClass({
-    render: function() {
-      return (
-  
-        <div className="login-page" style={{height: '100vh'}}>
+export class Login extends Component {
+
+  state = {
+    username: "",
+    password: ""
+  }
+
+  onChange = (e) => this.setState({[e.target.name]: e.target.value})
+
+  render() {
+    return (
+      <div className="login-page" style={{height: '100vh'}}>
           <div role="heading">
             <h1>Log In</h1>
           </div>
           <div className="form" role="form">
             <form className="login-form">
-              <input type="text" placeholder="username" />
-              <input type="text" placeholder="password" />
+              <input type="text" placeholder="Username" name="username" onChange={this.onChange}/>
+              <input type="text" placeholder="Password" name="password" onChange={this.onChange} />
               <button>login</button>
-              <p className="message">Not registered? <a href="create-login.html"><strong>Sign Up</strong></a></p>
+              <p className="message">Not registered? <Link to = "/create"/><strong>Sign Up</strong></p>
             </form>
           </div>
         </div>
-      );
-    }
-  });
+    )
+  }
+}
 
-  export default login;
+export default Login
+
+
+
+
+
+
+
+
+
+
